+++ ui/src/dashboards/components/DashboardHeader.js
@@ -1,6 +1,5 @@
 import React, {PropTypes} from 'react'
 import ReactTooltip from 'react-tooltip'
-import {Link} from 'react-router';
 
 import AutoRefreshDropdown from 'shared/components/AutoRefreshDropdown'
 import TimeRangeDropdown from 'shared/components/TimeRangeDropdown'
@@ -17,9 +16,9 @@ const DashboardHeader = ({
   handleChooseTimeRange,
   handleChooseAutoRefresh,
   handleClickPresentationButton,
-  sourceID,
   source,
   onAddCell,
+  onEditDashboard,
 }) => isHidden ? null : (
   <div className="page-header full-width">
     <div className="page-header__container">
@@ -47,11 +46,12 @@ const DashboardHeader = ({
               &nbsp;Add Cell
             </button> : null
         }
-        {sourceID ?
-          <Link className="btn btn-info btn-sm" to={`/sources/${sourceID}/dashboards/${dashboard && dashboard.id}/edit`} >
-            <span className="icon pencil" />
-            &nbsp;Edit
-          </Link> : null
+        {
+          dashboard ?
+            <button className="btn btn-info btn-sm" onClick={onEditDashboard}>
+              <span className="icon pencil" />
+              &nbsp;Edit
+            </button> : null
         }
         <div className="btn btn-info btn-sm" data-for="graph-tips-tooltip" data-tip="<p><code>Click + Drag</code> Zoom in (X or Y)</p><p><code>Shift + Click</code> Pan Graph Window</p><p><code>Double Click</code> Reset Graph Window</p>">
           <span className="icon heart"></span>
@@ -92,6 +92,7 @@ DashboardHeader.propTypes = {
   handleClickPresentationButton: func.isRequired,
   source: shape({}),
   onAddCell: func.isRequired,
+  onEditDashboard: func.isRequired,
 }
 
 export default DashboardHeader